cleaned up components and fixed sign in / sign out functionality
