+++ src/components/Layout.vue
@@ -16,12 +16,18 @@
     <v-toolbar :clipped-left="primaryDrawer.clipped" app absolute>
       <v-toolbar-side-icon
         v-if="primaryDrawer.type !== 'permanent'"
-        @click.stop="primaryDrawer.model = !primaryDrawer.model"
-      ></v-toolbar-side-icon>
-      <v-toolbar-title>Vuetify</v-toolbar-title>
+        @click.stop="primaryDrawer.model = !primaryDrawer.model">
+      </v-toolbar-side-icon>
+      <v-toolbar-title>Canoodle</v-toolbar-title>
+      <v-spacer></v-spacer>
+      <v-toolbar-items class="hidden-sm-and-down">
+        <v-btn flat to="/" class="fonted">Links</v-btn>
+        <v-btn flat v-if="!signedIn()" to="/signin" class="fonted">Sign In</v-btn>
+        <v-btn flat v-if="signedIn()" class="fonted" @click="signOut()">Sign Out</v-btn>
+      </v-toolbar-items>
     </v-toolbar>
     <v-content>
-      <SubjectList/>
+      <router-view/>
     </v-content>
     <v-footer :inset="footer.inset" app>
       <span class="px-3">&copy; {{ new Date().getFullYear() }}</span>
@@ -49,6 +55,34 @@ export default {
     footer: {
       inset: false
     }
-  })
+  }),
+  created () {
+    this.signedIn()
+  },
+  methods: {
+    setError (error, text) {
+      this.error = (error.response && error.response.data && error.response.data.error) || text
+    },
+    signedIn () {
+      this.signer = localStorage.signedIn
+      return localStorage.signedIn
+    },
+    signOut () {
+      localStorage.signedIn = false
+      this.signer = false
+      this.$http.secured.delete('/signin')
+        .then(response => {
+          delete localStorage.csrf
+          delete localStorage.signedIn
+          this.$forceUpdate()
+        })
+        .catch(error => this.setError(error, 'Cannot sign out'))
+    }
+  },
+  watch: {
+    signer: function (val) {
+      signedIn()
+    }
+  }
 }
 </script>
