+++ src/components/Signin.vue
@@ -4,7 +4,6 @@
       <h3 >Sign In</h3>
       <form @submit.prevent="signin">
         <div v-if="error">{{ error }}</div>
-
         <div v-if="!signedIn()">
           <label for="email" >E-mail Address</label>
           <input type="" v-model="email" id="email" placeholder="andy@web-crunch.com">
@@ -12,8 +11,6 @@
           <input type="password" v-model="password" id="password" placeholder="Password">
         </div>
         <button type="submit">Sign In</button>
-
-        <div><router-link to="/signup">Sign up</router-link></div>
       </form>
     </div>
   </div>
